move the index schema to a separate file
